Extract HtmlPolicy + SuccessHtmlRenderer from Core\Utils (#563 Sprint 5 phase 2, B1)
Move the two HTML concerns out of the Core\Utils hub:

- get_allowed_html_tags() -> FreeFormCertificate\Core\HtmlPolicy (the shared
  wp_kses allow-list used by Frontend / Email / PDF sanitization).
- generate_success_html() -> FreeFormCertificate\Frontend\Submission\SuccessHtmlRenderer,
  next to its sole caller SuccessResponder — it is a frontend view concern,
  not a general-purpose utility.

All call sites repointed (no back-compat wrappers); the DataSanitizer test
alias-mock now targets HtmlPolicy, and the dedicated allowed-tags /
success-html tests move to HtmlPolicyTest / the repointed SuccessHtmlTest.

After this commit Core\Utils holds only four generic primitives
(format_bytes, truncate, get_day_of_week_number, sanitize_username_slug) —
the B1 end state of "keep only generic primitives". Behavior-preserving.
