+++ includes/core/class-ffc-utils.php
@@ -25,121 +25,6 @@ if ( ! defined( 'ABSPATH' ) ) {
 class Utils {
 
 	/**
-	 * Returns the list of allowed HTML tags and attributes.
-	 * Centralized here so Frontend, Email, and PDF Generator use the same validation rules.
-	 *
-	 * @return array<string, array<string, array<never, never>>> Allowed HTML tags with their attributes
-	 */
-	public static function get_allowed_html_tags(): array {
-		$allowed = array(
-			'b'      => array(),
-			'strong' => array(),
-			'i'      => array(),
-			'em'     => array(),
-			'u'      => array(),
-			'br'     => array(),
-			'hr'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'p'      => array(
-				'style' => array(),
-				'class' => array(),
-				'align' => array(),
-			),
-			'span'   => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'div'    => array(
-				'style' => array(),
-				'class' => array(),
-				'id'    => array(),
-			),
-			'font'   => array(
-				'color' => array(),
-				'size'  => array(),
-				'face'  => array(),
-			),
-			'img'    => array(
-				'src'    => array(),
-				'alt'    => array(),
-				'style'  => array(),
-				'width'  => array(),
-				'height' => array(),
-			),
-			// Table tags (essential for signature alignment).
-			'table'  => array(
-				'style'       => array(),
-				'class'       => array(),
-				'width'       => array(),
-				'border'      => array(),
-				'cellpadding' => array(),
-				'cellspacing' => array(),
-				'role'        => array(),
-			),
-			'tr'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'td'     => array(
-				'style'   => array(),
-				'width'   => array(),
-				'colspan' => array(),
-				'rowspan' => array(),
-				'align'   => array(),
-				'valign'  => array(),
-			),
-			'th'     => array(
-				'style'   => array(),
-				'width'   => array(),
-				'colspan' => array(),
-				'rowspan' => array(),
-				'align'   => array(),
-				'valign'  => array(),
-			),
-			// Headings.
-			'h1'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'h2'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'h3'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'h4'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-
-			// Lists (useful for syllabus content on the back or body).
-			'ul'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'ol'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-			'li'     => array(
-				'style' => array(),
-				'class' => array(),
-			),
-		);
-
-		/**
-		 * Allows developers to filter or add new tags
-		 * without modifying the plugin core.
-		 */
-        // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.NonPrefixedHooknameFound -- ffcertificate is the plugin prefix
-		return apply_filters( 'ffcertificate_allowed_html_tags', $allowed );
-	}
-
-	/**
 	 * Convert bytes to human-readable format
 	 *
 	 * @param int $bytes Number of bytes.
@@ -232,66 +117,5 @@ class Utils {
 		$slug = preg_replace( '/[^a-z0-9._-]/', '', $slug ) ?? '';
 		$slug = preg_replace( '/[-_.]+/', '.', $slug ) ?? '';
 		return trim( $slug, '.' );
-	}
-
-	/**
-	 * Generate success HTML response for frontend form submission
-	 *
-	 * Generate success html.
-	 *
-	 * Generate success html.
-	 *
-	 * Generate success html.
-	 *
-	 * Generate success html.
-	 *
-	 * Generate success html.
-	 *
-	 * @since 2.9.16
-	 * @param array<string, mixed> $submission_data Submission data.
-	 * @param int                  $form_id Form ID.
-	 * @param int|string           $submission_date Submission date — unix UTC int (since 6.6.0, reprint flow)
-	 *                                              or MySQL `Y-m-d H:i:s` string (fresh submission via `current_time('mysql')`).
-	 *                                              DateFormatter::format_datetime accepts both.
-	 * @param string               $success_message Success message.
-	 * @param int                  $submission_id Submission ID (used to surface the magic link in the success card; 0 to skip).
-	 * @param object|null          $submission_handler Handler that knows how to ensure/load the submission's magic token.
-	 * @return string HTML content
-	 */
-	public static function generate_success_html( array $submission_data, int $form_id, int|string $submission_date, string $success_message = '', int $submission_id = 0, ?object $submission_handler = null ): string {
-		// Get form configuration.
-		$form_config = get_post_meta( $form_id, '_ffc_form_config', true );
-		if ( ! is_array( $form_config ) ) {
-			$form_config = array();
-		}
-
-		// Get form title.
-		$form_post  = get_post( $form_id );
-		$form_title = $form_post ? $form_post->post_title : __( 'Certificate', 'ffcertificate' );
-
-		// Default success message.
-		if ( empty( $success_message ) ) {
-			$success_message = isset( $form_config['success_message'] ) && ! empty( $form_config['success_message'] )
-				? $form_config['success_message']
-				: __( 'Success! Your certificate has been generated.', 'ffcertificate' );
-		}
-
-		$date_formatted = \FreeFormCertificate\Core\DateFormatter::format_datetime( $submission_date );
-
-		// Auth code (formatted for display with certificate prefix).
-		$auth_code = isset( $submission_data['auth_code'] ) ? DocumentFormatter::format_auth_code( $submission_data['auth_code'], DocumentFormatter::PREFIX_CERTIFICATE ) : '';
-
-		// Magic link — survives the tab close, so the user can come back
-		// later from a different device and re-issue the certificate.
-		$magic_link = '';
-		if ( $submission_id > 0 && $submission_handler && class_exists( '\FreeFormCertificate\Generators\MagicLinkHelper' ) ) {
-			$magic_link = \FreeFormCertificate\Generators\MagicLinkHelper::get_submission_magic_link( $submission_id, $submission_handler );
-		}
-
-		// Load template.
-		ob_start();
-		include FFC_PLUGIN_DIR . 'templates/submission-success.php';
-		$rendered = ob_get_clean();
-		return $rendered ? $rendered : '';
 	}
 }
